refactor(webapp): share the ownership-transfer gate and the layout test fixture
Addresses two CodeRabbit findings on #2861.

- add useCanTransferOwnership, so the two surfaces that gate on "workspace
  owner or Shelf staff admin" (the team-page button and the general-settings
  card) name the rule once. They live on different pages and must agree, or a
  user sees a live link on one and "only the owner can transfer" on the other.
  The staff-admin term is load-bearing: the admin dashboard renders the card
  for Shelf staff who are not members of the workspace, so isOwner is false
  there. The gate stays UI-cosmetic — transferOwnership re-checks both
  server-side
- move the duplicated _layout loader payload into a shared factory,
  createLayoutLoaderData in test/factories, keeping the nullable-owner knob.
  The two copies were identical apart from that one parameter, and this PR had
  just edited both, which is how they drift

No behavior change.
