+++ apps/webapp/app/components/settings/transfer-ownership-button.tsx
@@ -1,6 +1,5 @@
+import { useCanTransferOwnership } from "~/hooks/use-can-transfer-ownership";
 import { useCurrentOrganization } from "~/hooks/use-current-organization";
-import { useIsShelfAdmin } from "~/hooks/use-is-shelf-admin";
-import { useUserRoleHelper } from "~/hooks/user-user-role-helper";
 import { Button } from "../shared/button";
 
 /**
@@ -12,9 +11,9 @@ import { Button } from "../shared/button";
  * to "Import Users" / "Invite a user" so it reads as a page-level action, not
  * a row action.
  *
- * - Workspace owner (and Shelf staff admins, via {@link useIsShelfAdmin} —
- *   the same gate TransferOwnershipCard uses) get a link to the transfer
- *   section.
+ * - Whoever may run the transfer (workspace owner or Shelf staff admin, via
+ *   {@link useCanTransferOwnership} — the same gate TransferOwnershipCard
+ *   uses) gets a link to the transfer section.
  * - Everyone else gets a disabled button whose hover reason names the owner,
  *   so admins learn who can run the transfer instead of the feature being
  *   invisible to them.
@@ -22,8 +21,7 @@ import { Button } from "../shared/button";
  * @see {@link file://./transfer-ownership-card.tsx}
  */
 export default function TransferOwnershipButton() {
-  const { isOwner } = useUserRoleHelper();
-  const isShelfAdmin = useIsShelfAdmin();
+  const canTransferOwnership = useCanTransferOwnership();
   const currentOrganization = useCurrentOrganization();
 
   /** Shown to non-owners so they know who to ask */
@@ -31,7 +29,7 @@ export default function TransferOwnershipButton() {
 
   // No width/margin classes on either branch: the actions row that renders
   // this owns the layout (see settings.team.users).
-  if (isOwner || isShelfAdmin) {
+  if (canTransferOwnership) {
     return (
       <Button to="/settings/general#transfer-ownership" variant="secondary">
         <span className="whitespace-nowrap">Transfer ownership</span>
