+++ apps/webapp/test/factories/index.ts
@@ -9,3 +9,4 @@ export * from "./asset";
 export * from "./organization";
 export * from "./assetModel";
 export * from "./activityEvent";
+export * from "./layoutLoaderData";
